+++ src/options.js
@@ -68,6 +68,7 @@ defaultOptions.ui.bootstrap2 = false;
 defaultOptions.ui.bootstrap4 = false;
 defaultOptions.ui.showProgressBar = true;
 defaultOptions.ui.showPopover = false;
+defaultOptions.ui.popoverPlacement = "bottom";
 defaultOptions.ui.showStatus = false;
 defaultOptions.ui.spanError = function (options, key) {
     "use strict";